chore: move to xds inspect
